feat: add join and leave playlist feature
The other users should be able to join and leave a public playlist using
its invite code. And the owner should be able to delete its guests.

Playlist and User model has now a many-to-many relationship with a
relationship table named as userPlaylist. Routes to leave, join, list
guests, remove guests are added.