feat(frontend): implement auth context to all pages
Implement Auth Context functions to login and signup pages, as well as
playlist page, redirecting to login page when trying to enter a page
that requires authentication and user is not authenticated.

Add a Layout page and implement logout button and an Outlet for every
route. Remove Outlet from Home page and adapt to receive childrens such
as login and signup pages.